puta madre

+++ KafeYana.Api/KafeYana.Domain/Dtos/VariacionesDtos/DtoAjusteCU.cs
@@ -12,6 +12,6 @@ namespace KafeYana.Application.Dtos.VariacionesDtos
 
         public int? Id_InsumoNuevo { get; set; }
 
-        public required int Cantidad { get; set; } 
+        public required decimal Cantidad { get; set; } 
     }
 }
